fix(distributed-locks): address lease monitor review findings
